apx-app/pyproject.toml
[project]
name = "sobeys-scctower"
dynamic = ["version"]
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "pydantic-settings>=2.11.0",
    "uvicorn>=0.37.0",
    "databricks-sdk>=0.74.0",
    "sqlmodel>=0.0.27",
    "psycopg[binary,pool]>=3.2.11",
    "openai>=2.21.0",
    "pyyaml>=6.0.3",
    "httpx>=0.28.1",
]

[dependency-groups]
dev = [
    "ty>=0.0.12", "apx==0.2.7",
]

[tool.apx.metadata]
app-name = "sobeys-scctower"
app-slug = "sobeys_scctower"
app-entrypoint = "sobeys_scctower.backend.app:app"
api-prefix = "/api"
metadata-path = "src/sobeys_scctower/_metadata.py"

[tool.apx.ui]
root = "src/sobeys_scctower/ui"

[tool.apx.ui.registries]
"@animate-ui" = "https://animate-ui.com/r/{name}.json"
"@ai-elements" = "https://registry.ai-sdk.dev/{name}.json"
"@svgl" = "https://svgl.app/r/{name}.json"

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.uv-dynamic-versioning]
vcs = "git"
fallback-version = "0.0.0"

[tool.uv]

[[tool.uv.index]]
name = "apx-index"
url = "https://databricks-solutions.github.io/apx/simple"

[tool.uv.sources]

[tool.uv.sources.apx]
index = "apx-index"

[tool.hatch.build.hooks.version]
path = "src/sobeys_scctower/_version.py"
template = '''
version = "{version}"
'''

[tool.hatch.build]
artifacts = ["src/sobeys_scctower/__dist__", "src/sobeys_scctower/_metadata.py"]
exclude = ["src/sobeys_scctower/ui"]
